Add Embed Trait overall
