+++ src/Builder/Pdf/EmbedPdfBuilder.php
@@ -7,6 +7,7 @@ use Sensiolabs\GotenbergBundle\Builder\Attributes\NormalizeGotenbergPayload;
 use Sensiolabs\GotenbergBundle\Builder\Attributes\WithBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\Dependencies\AssetBaseDirFormatterAwareTrait;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\DownloadFromTrait;
+use Sensiolabs\GotenbergBundle\Builder\Behaviors\EmbedTrait;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\WebhookTrait;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
 use Sensiolabs\GotenbergBundle\Exception\MissingRequiredFieldException;
@@ -21,6 +22,7 @@ final class EmbedPdfBuilder extends AbstractBuilder
 {
     use AssetBaseDirFormatterAwareTrait;
     use DownloadFromTrait;
+    use EmbedTrait;
     use WebhookTrait;
 
     public const ENDPOINT = '/forms/pdfengines/embed';
@@ -47,31 +49,6 @@ final class EmbedPdfBuilder extends AbstractBuilder
         }
 
         $this->getBodyBag()->set('files', $files ?? null);
-
-        return $this;
-    }
-
-    /**
-     * Add file to embed.
-     *
-     * As assets files, by default the files to embed are fetch in the assets folder
-     * of your application. For more information about path resolution go to
-     * assets documentation.
-     *
-     * @see https://gotenberg.dev/docs/routes#embed-files-route
-     *
-     * @example embeds('document.xml','document_2.json')
-     */
-    public function embeds(string|\Stringable ...$paths): self
-    {
-        foreach ($paths as $path) {
-            $path = (string) $path;
-
-            $info = new \SplFileInfo($this->getAssetBaseDirFormatter()->resolve($path));
-            $files[$path] = $info;
-        }
-
-        $this->getBodyBag()->set('embeds', $files ?? null);
 
         return $this;
     }
