+++ src/Builder/Pdf/LibreOfficePdfBuilder.php
@@ -6,6 +6,7 @@ use Sensiolabs\GotenbergBundle\Builder\AbstractBuilder;
 use Sensiolabs\GotenbergBundle\Builder\Attributes\NormalizeGotenbergPayload;
 use Sensiolabs\GotenbergBundle\Builder\Attributes\WithBuilderConfiguration;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\Dependencies\AssetBaseDirFormatterAwareTrait;
+use Sensiolabs\GotenbergBundle\Builder\Behaviors\EmbedTrait;
 use Sensiolabs\GotenbergBundle\Builder\Behaviors\LibreOfficeTrait;
 use Sensiolabs\GotenbergBundle\Builder\Util\NormalizerFactory;
 use Sensiolabs\GotenbergBundle\Builder\Util\ValidatorFactory;
@@ -22,6 +23,7 @@ use Sensiolabs\GotenbergBundle\Exception\MissingRequiredFieldException;
 final class LibreOfficePdfBuilder extends AbstractBuilder
 {
     use AssetBaseDirFormatterAwareTrait;
+    use EmbedTrait;
     use LibreOfficeTrait;
 
     public const ENDPOINT = '/forms/libreoffice/convert';
@@ -55,35 +57,6 @@ final class LibreOfficePdfBuilder extends AbstractBuilder
         }
 
         $this->getBodyBag()->set('files', $files ?? null);
-
-        return $this;
-    }
-
-    /**
-     * Add file to embed.
-     *
-     * As assets files, by default the files to embed are fetch in the assets folder
-     * of your application. For more information about path resolution go to
-     * assets documentation.
-     *
-     * @see https://gotenberg.dev/docs/routes#embeds-libreoffice
-     *
-     * @example embeds('document.xml','document_2.json')
-     */
-    public function embeds(string|\Stringable ...$paths): self
-    {
-        $this->introducedIn('8.25');
-
-        foreach ($paths as $path) {
-            $path = (string) $path;
-
-            $info = new \SplFileInfo($this->getAssetBaseDirFormatter()->resolve($path));
-            ValidatorFactory::filesExtension([$info], ['xml', 'json']);
-
-            $files[$path] = $info;
-        }
-
-        $this->getBodyBag()->set('embeds', $files ?? null);
 
         return $this;
     }
